+++ app/components/homePageComponents/SearchBar.tsx
